requirements.txt
Django>=5.0,<6
python-dotenv>=1.0
hyperliquid-python-sdk>=0.20.0
eth-account>=0.11.0
eth-keys>=0.5.0
eth-utils>=2.0.0
